Refactor and bug correction. LSMOP has been renamed to AbstractLSMOP and fill up all problem information; as a consequence the constructor of every problem gets significantly reduced. A new class AbstractLSMOP1_4 is introduced to implement common behavior for problems 1_4 Problems LSMOP1 to 3 have bee refactored according to these changes.
