+++ jmetal-problem/src/main/java/org/uma/jmetal/problem/multiobjective/lsmop/LSMOP2.java
@@ -1,10 +1,8 @@
 package org.uma.jmetal.problem.multiobjective.lsmop;
 
-import org.uma.jmetal.problem.doubleproblem.impl.AbstractDoubleProblem;
 import org.uma.jmetal.problem.multiobjective.lsmop.functions.Function;
 import org.uma.jmetal.problem.multiobjective.lsmop.functions.Griewank;
 import org.uma.jmetal.problem.multiobjective.lsmop.functions.Schwefel;
-import org.uma.jmetal.problem.multiobjective.lsmop.functions.Sphere;
 import org.uma.jmetal.solution.doublesolution.DoubleSolution;
 import org.uma.jmetal.util.errorchecking.JMetalException;
 
@@ -15,7 +13,7 @@ import java.util.List;
  * Class representing problem LSMOP2
  */
 
-public class LSMOP2 extends LSMOP {
+public class LSMOP2 extends AbstractLSMOP1_4 {
 
 
     /**
@@ -37,37 +35,16 @@ public class LSMOP2 extends LSMOP {
     public LSMOP2(int nk, int numberOfVariables, int numberOfObjectives) throws JMetalException {
         super(nk,numberOfVariables,numberOfObjectives);
         setName("LSMOP2");
-
-        List<Double> lowerLimit = new ArrayList<Double>(getNumberOfVariables()) ;
-        List<Double> upperLimit = new ArrayList<Double>(getNumberOfVariables()) ;
-
-        for (int i = 0; i < getNumberOfObjectives()-1; i++) {
-            lowerLimit.add(0.0);
-            upperLimit.add(1.0);
-        }
-
-        for (int i = getNumberOfObjectives()-1; i < getNumberOfVariables(); i++) {
-            lowerLimit.add(0.0);
-            upperLimit.add(10.0);
-        }
-
-        setVariableBounds(lowerLimit, upperLimit);
     }
 
+    @Override
+    protected Function getOddFunction() {
+        return new Griewank();
+    }
 
-    @java.lang.Override
-    public DoubleSolution evaluate(DoubleSolution solution) {
-        List<Double> variables = new ArrayList<>(getNumberOfVariables());
-
-        for (int i = 0; i < getNumberOfVariables(); i++) {
-            variables.add(solution.variables().get(i));
-        }
-        List<Double> y = evaluate(variables,new Griewank(),new Schwefel());
-
-        for (int i = 0; i < getNumberOfObjectives(); i++) {
-            solution.objectives()[i] = y.get(i);
-        }
-        return solution;
+    @Override
+    protected Function getEvenFunction() {
+        return new Schwefel();
     }
 
 }
