+++ jmetal-problem/src/main/java/org/uma/jmetal/problem/multiobjective/lsmop/LSMOP3.java
@@ -1,10 +1,8 @@
 package org.uma.jmetal.problem.multiobjective.lsmop;
 
-import org.uma.jmetal.problem.doubleproblem.impl.AbstractDoubleProblem;
 import org.uma.jmetal.problem.multiobjective.lsmop.functions.Function;
 import org.uma.jmetal.problem.multiobjective.lsmop.functions.Rastrigin;
 import org.uma.jmetal.problem.multiobjective.lsmop.functions.Rosenbrock;
-import org.uma.jmetal.problem.multiobjective.lsmop.functions.Sphere;
 import org.uma.jmetal.solution.doublesolution.DoubleSolution;
 import org.uma.jmetal.util.errorchecking.JMetalException;
 
@@ -15,11 +13,11 @@ import java.util.List;
  * Class representing problem LSMOP3
  */
 
-public class LSMOP3 extends LSMOP {
+public class LSMOP3 extends AbstractLSMOP1_4 {
 
 
     /**
-     * Creates a default LSMOP1 problem (7 variables and 3 objectives)
+     * Creates a default LSMOP3 problem (7 variables and 3 objectives)
      */
     public LSMOP3() {
         this(5, 300, 3);
@@ -37,37 +35,16 @@ public class LSMOP3 extends LSMOP {
     public LSMOP3(int nk, int numberOfVariables, int numberOfObjectives) throws JMetalException {
         super(nk,numberOfVariables,numberOfObjectives);
         setName("LSMOP3");
-
-        List<Double> lowerLimit = new ArrayList<Double>(getNumberOfVariables()) ;
-        List<Double> upperLimit = new ArrayList<Double>(getNumberOfVariables()) ;
-
-        for (int i = 0; i < getNumberOfObjectives()-1; i++) {
-            lowerLimit.add(0.0);
-            upperLimit.add(1.0);
-        }
-
-        for (int i = getNumberOfObjectives()-1; i < getNumberOfVariables(); i++) {
-            lowerLimit.add(0.0);
-            upperLimit.add(10.0);
-        }
-
-        setVariableBounds(lowerLimit, upperLimit);
     }
 
+    @Override
+    protected Function getOddFunction() {
+        return new Rastrigin();
+    }
 
-    @java.lang.Override
-    public DoubleSolution evaluate(DoubleSolution solution) {
-        List<Double> variables = new ArrayList<>(getNumberOfVariables());
-
-        for (int i = 0; i < getNumberOfVariables(); i++) {
-            variables.add(solution.variables().get(i));
-        }
-        List<Double> y = evaluate(variables,new Rastrigin(),new Rosenbrock());
-
-        for (int i = 0; i < getNumberOfObjectives(); i++) {
-            solution.objectives()[i] = y.get(i);
-        }
-        return solution;
+    @Override
+    protected Function getEvenFunction() {
+        return new Rosenbrock();
     }
 
 }
